desafio atualizar estdo

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -2,7 +2,10 @@ package com.algaworks.algafood.infrastructure.repository;
 
 import com.algaworks.algafood.domain.model.Estado;
 import com.algaworks.algafood.domain.repository.EstadoRepository;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -24,14 +27,19 @@ public class EstadoRepositoryImpl implements EstadoRepository {
         return entityManager.find(Estado.class, id);
     }
 
+    @Transactional
     @Override
     public Estado adicionar(Estado estado) {
         return entityManager.merge(estado);
     }
 
+    @Transactional
     @Override
-    public void remover(Estado estado) {
-        estado = buscar(estado.getId());
+    public void remover(Long estadoId) {
+        Estado estado = buscar(estadoId);
+        if (estado == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(estado);
     }
 }
